Add basic options to the runner
