+++ runner/src/main/scala/hedgehog/runner/Properties.scala
@@ -1,7 +1,8 @@
 package hedgehog.runner
 
-import hedgehog._
-import hedgehog.core._
+import hedgehog.Property
+import hedgehog.core.Seed
+import scopt.OParser
 
 abstract class Properties {
 
@@ -9,90 +10,54 @@ abstract class Properties {
 
   /** Allows the implementing test to be run separately without SBT */
   def main(args: Array[String]): Unit = {
-    val config = PropertyConfig.default
-    val seed = Seed.fromTime()
+    OParser.parse(optionParser, args, Config.default) match {
+      case Some(config) =>
+        val tests = filteredTests(config)
+        if (tests.isEmpty) {
+          println("No tests to run.")
+        } else {
+          runTests(tests, config)
+        }
+      case _ =>
+        System.err.println()
+        sys.error("Failed to parse test arguments.")
+    }
+  }
+
+  private def optionParser: OParser[_, Config] = {
+    val builder = OParser.builder[Config]
+    import builder._
+    OParser.sequence(
+      programName(this.getClass.getSimpleName.replaceAll("""\$$""", "")),
+      opt[String]('n', "name")
+        .action((name, config) => config.copy(testNames = name :: config.testNames))
+        .text("Only runs tests with the given name. Specify multiple times to run multiple tests.")
+        .unbounded(),
+      opt[Long]('s', "seed")
+        .action((seed, config) => config.copy(seed = seed))
+        .text("Sets the seed value to use."),
+      help('h', "help")
+    )
+  }
+
+  private def filteredTests(config: Config): List[Test] =
+    config.testNames match {
+      case Nil => tests
+      case names =>
+        for {
+          test <- tests
+          name <- names
+          if test.name == name
+        } yield test
+    }
+
+  def runTests(tests: List[Test], config: Config): Unit = {
+    val propertyConfig = config.propertyConfig
+    val seed = config.seed
+    println(s"Running tests (seed = $seed):")
     tests.foreach(t => {
-      val report = Property.check(t.withConfig(config), t.result, seed)
+      val report = Property.check(t.withConfig(propertyConfig), t.result, Seed.fromLong(seed))
       println(Test.renderReport(this.getClass.getName, t, report, ansiCodesSupported = true))
     })
   }
-}
-
-class Test(
-    val name: String
-  , val withConfig: PropertyConfig => PropertyConfig
-  , val result: Property
-  ) {
-
-  def config(f: PropertyConfig => PropertyConfig): Test =
-    new Test(name, c => f(withConfig(c)), result)
-
-  def withTests(count: SuccessCount): Test =
-    config(_.copy(testLimit = count))
-
-  def noShrinking: Test =
-    config(_.copy(shrinkLimit = ShrinkLimit(0)))
-}
-
-object Test {
-
-  /** Wrap the actual constructor so we can catch any exceptions thrown */
-  def apply(name: String, result: => Property): Test =
-    try {
-      new Test(name, identity, result)
-    } catch {
-      case e: Exception =>
-        new Test(name, identity, Property.error(e))
-    }
-
-  def renderReport(className: String, t: Test, report: Report, ansiCodesSupported: Boolean): String = {
-    def render(ok: Boolean, msg: String, extraS: List[String]): String = {
-      val name = className + "." + t.name
-      val sym = if (ok) "+" else "-"
-      val colour = if (ok) Console.GREEN else Console.RED
-      val extra = if (extraS.isEmpty) "" else "\n" + extraS.map(s => "> " + s).mkString("\n")
-      if(ansiCodesSupported) {
-        s"$colour$sym${Console.RESET} $name: $msg$extra"
-      } else {
-        s"$sym $name: $msg$extra"
-      }
-    }
-
-    val coverage = renderCoverage(report.coverage, report.tests)
-    report.status match {
-      case Failed(shrinks, log) =>
-        render(false, s"Falsified after ${report.tests.value} passed tests", log.map(renderLog) ++ coverage)
-      case GaveUp =>
-        render(false, s"Gave up after only ${report.tests.value} passed test. " +
-          s"${report.discards.value} were discarded", coverage)
-      case OK =>
-        render(true, s"OK, passed ${report.tests.value} tests", coverage)
-    }
-  }
-
-  def renderLog(log: Log): String =
-    log match {
-      case ForAll(name, value) =>
-        s"${name.value}: $value"
-      case Info(value) =>
-        value
-      case Error(e) =>
-        val sw = new java.io.StringWriter()
-        e.printStackTrace(new java.io.PrintWriter(sw))
-        sw.toString
-    }
-
-  def renderCoverage(coverage: Coverage[CoverCount], tests: SuccessCount): List[String] =
-    coverage.labels.values.toList
-      .sortBy(_.annotation.percentage(tests).toDouble.toInt * -1)
-      .map(l => {
-        List(
-          List(l.annotation.percentage(tests).toDouble.toInt.toString + "%")
-        , List(l.name.render)
-        , if (l.minimum.toDouble > 0) List(
-            l.minimum.toDouble.toInt.toString + "%"
-          , if (Label.covered(l, tests)) "✓" else "✗"
-          ) else Nil
-        ).flatten.mkString(" ")
-      })
 }
